swap : and #
swap the symbol `:` ẅith `#` in order to better conform a design
principle, where when you have a SFB (same-finger bigrams), like
in `::` you avoid to use the pinky finger.

Also, in this new no position, `:` became a analogous position with `;`.
Both with same finger and row, but with different hands.